fixed AssetHeader struct

+++ lib/NanoUI/core/types.h
@@ -51,8 +51,8 @@ struct Region
 
 struct __attribute__((packed)) AssetHeader{
     char magic[4];
-    uint32_t version;
-    uint32_t count;
+    uint16_t version;
+    uint16_t count;
     uint32_t table_offset;
     uint32_t data_offset;
     uint32_t reserved;
